Re-factor HDFFile.
