+++ src/HDFFile.h
@@ -17,114 +17,49 @@
 #include <h5cpp/hdf5.hpp>
 #include <string>
 #include <vector>
+#include "StreamHDFInfo.h"
 
 namespace FileWriter {
 
-// POD
-struct StreamHDFInfo {
-  std::string HDFParentName;
-  std::string ConfigStream;
-  bool InitialisedOk = false;
-};
-
-class HDFFile {
+class HDFFileBase {
 public:
-  HDFFile();
+  virtual ~HDFFileBase() = default;
+  virtual void flush();
 
-  ~HDFFile();
+  auto hdfGroup() const {
+    return H5File.root();
+  }
 
-  void init(const std::string &Filename, nlohmann::json const &NexusStructure,
-            std::vector<StreamHDFInfo> &StreamHDFInfo);
-
+protected:
+  auto& hdfFile() {
+    return H5File;
+  }
   void init(const std::string &NexusStructure,
-            std::vector<StreamHDFInfo> &StreamHDFInfo);
+                std::vector<StreamHDFInfo> &StreamHDFInfo);
 
   void init(const nlohmann::json &NexusStructure,
-            std::vector<StreamHDFInfo> &StreamHDFInfo);
+                std::vector<StreamHDFInfo> &StreamHDFInfo);
 
-  void reopen(std::string const &Filename);
-
-  void flush();
-  void close();
-  void finalize();
-
-  hdf5::file::File H5File;
-  hdf5::node::Group RootGroup;
-
-private:
-  std::string Filename;
-  nlohmann::json NexusStructure;
-
-  using CLOCK = std::chrono::steady_clock;
   SharedLogger Logger = getLogger();
+private:
+  hdf5::file::File H5File;
 };
 
-bool findType(nlohmann::basic_json<> Attribute, std::string &DType);
-std::string h5VersionStringLinked();
-void writeAttributes(hdf5::node::Node const &Node, nlohmann::json const *Value,
-                     SharedLogger const &Logger);
+class HDFFile : public HDFFileBase {
+public:
+  HDFFile(std::string const &FileName, nlohmann::json const &NexusStructure,
+          std::vector<StreamHDFInfo> &StreamHDFInfo);
+  virtual ~HDFFile();
+private:
+  void createFileInRegularMode();
+  void openFileInRegularMode();
+  void openFileInSWMRMode();
+  void closeFile();
+  void addLinks();
 
-void writeStringAttribute(hdf5::node::Node const &Node, std::string const &Name,
-                          std::string const &Value);
+  std::string H5FileName;
+  nlohmann::json StoredNexusStructure;
+};
 
-void checkHDFVersion(SharedLogger const &Logger);
-std::string H5VersionStringHeadersCompileTime();
 
-void createHDFStructures(
-    const nlohmann::json *Value, hdf5::node::Group const &Parent,
-    uint16_t Level,
-    hdf5::property::LinkCreationList const &LinkCreationPropertyList,
-    hdf5::datatype::String const &FixedStringHDFType,
-    std::vector<StreamHDFInfo> &HDFStreamInfo, std::deque<std::string> &Path,
-    SharedLogger const &Logger);
-
-void writeHDFISO8601AttributeCurrentTime(hdf5::node::Node const &Node,
-                                         const std::string &Name,
-                                         SharedLogger const &Logger);
-
-void writeAttributesIfPresent(hdf5::node::Node const &Node,
-                              nlohmann::json const &Values,
-                              SharedLogger const &Logger);
-
-std::vector<std::string> populateStrings(const nlohmann::json *Values,
-                                         hssize_t GoalSize);
-
-void writeStringDataset(
-    hdf5::node::Group const &Parent, const std::string &Name,
-    hdf5::property::DatasetCreationList &DatasetCreationList,
-    hdf5::dataspace::Dataspace &Dataspace, nlohmann::json const &Values);
-
-void writeFixedSizeStringDataset(
-    hdf5::node::Group const &Parent, const std::string &Name,
-    hdf5::property::DatasetCreationList &DatasetCreationList,
-    hdf5::dataspace::Dataspace &Dataspace, hsize_t ElementSize,
-    const nlohmann::json *Values, SharedLogger const &Logger);
-
-void writeGenericDataset(const std::string &DataType,
-                         hdf5::node::Group const &Parent,
-                         const std::string &Name,
-                         const std::vector<hsize_t> &Sizes,
-                         const std::vector<hsize_t> &Max, hsize_t ElementSize,
-                         const nlohmann::json *Values,
-                         SharedLogger const &Logger);
-
-void writeDataset(hdf5::node::Group const &Parent, const nlohmann::json *Values,
-                  SharedLogger const &Logger);
-
-void writeObjectOfAttributes(hdf5::node::Node const &Node,
-                             const nlohmann::json &Values);
-
-void writeArrayOfAttributes(hdf5::node::Node const &Node,
-                            const nlohmann::json &ValuesJson,
-                            SharedLogger const &Logger);
-
-void writeScalarAttribute(hdf5::node::Node const &Node, const std::string &Name,
-                          const nlohmann::json &Values);
-
-void writeAttrOfSpecifiedType(std::string const &DType,
-                              hdf5::node::Node const &Node,
-                              std::string const &Name, uint32_t StringSize,
-                              hdf5::datatype::CharacterEncoding Encoding,
-                              nlohmann::json const &Values,
-                              SharedLogger const &Logger);
 } // namespace FileWriter
