+++ src/StreamHDFInfo.h
@@ -1,0 +1,18 @@
+// SPDX-License-Identifier: BSD-2-Clause
+//
+// This code has been produced by the European Spallation Source
+// and its partner institutes under the BSD 2 Clause License.
+//
+// See LICENSE.md at the top level for license information.
+//
+// Screaming Udder!                              https://esss.se
+
+#pragma once
+
+#include <string>
+
+struct StreamHDFInfo {
+  std::string HDFParentName;
+  std::string ConfigStream;
+  bool InitialisedOk = false;
+};
