+++ src/tests/WriterModule/ns10_WriterTests.cpp
@@ -78,16 +78,14 @@ public:
     registerSchema();
 
     File = HDFFileTestHelper::createInMemoryTestFile(TestFileName);
-    RootGroup = File.H5File.root();
+    RootGroup = File->hdfGroup();
     UsedGroup = RootGroup.create_group(NXLogGroup);
     setExtractorModule<AccessMessageMetadata::ns10_Extractor>("ns10");
   };
 
-  void TearDown() override { File.close(); };
-
   std::string TestFileName{"SomeTestFile.hdf5"};
   std::string NXLogGroup{"SomeParentName"};
-  FileWriter::HDFFile File;
+  std::unique_ptr<HDFFileTestHelper::DebugHDFFile> File;
   hdf5::node::Group RootGroup;
   hdf5::node::Group UsedGroup;
   hdf5::file::MemoryDriver MemoryDriver;
