+++ src/tests/WriterModule/senv_WriterTests.cpp
@@ -43,15 +43,14 @@ class FastSampleEnvironmentWriter : public ::testing::Test {
 public:
   void SetUp() override {
     File = HDFFileTestHelper::createInMemoryTestFile(TestFileName);
-    RootGroup = File.H5File.root();
+    RootGroup = File->hdfGroup();
     UsedGroup = RootGroup.create_group(NXLogGroup);
     setExtractorModule<AccessMessageMetadata::senv_Extractor>("senv");
   };
 
-  void TearDown() override { File.close(); };
   std::string TestFileName{"SomeTestFile.hdf5"};
   std::string NXLogGroup{"SomeParentName"};
-  FileWriter::HDFFile File;
+  std::unique_ptr<HDFFileTestHelper::DebugHDFFile> File;
   hdf5::node::Group RootGroup;
   hdf5::node::Group UsedGroup;
 };
