+++ src/tests/WriterModule/tdct_WriterTests.cpp
@@ -39,15 +39,14 @@ class ChopperTimeStampWriter : public ::testing::Test {
 public:
   void SetUp() override {
     File = HDFFileTestHelper::createInMemoryTestFile(TestFileName);
-    RootGroup = File.H5File.root();
+    RootGroup = File->hdfGroup();
     UsedGroup = RootGroup.create_group(NXLogGroup);
     setExtractorModule<AccessMessageMetadata::tdct_Extractor>("tdct");
   };
 
-  void TearDown() override { File.close(); };
   std::string TestFileName{"SomeTestFile.hdf5"};
   std::string NXLogGroup{"SomeParentName"};
-  FileWriter::HDFFile File;
+  std::unique_ptr<HDFFileTestHelper::DebugHDFFile> File;
   hdf5::node::Group RootGroup;
   hdf5::node::Group UsedGroup;
 };
