+++ src/tests/helpers/HDFFileTestHelper.h
@@ -15,7 +15,22 @@
 
 namespace HDFFileTestHelper {
 
-FileWriter::HDFFile createInMemoryTestFile(const std::string &Filename,
+class DebugHDFFile : public FileWriter::HDFFileBase {
+public:
+  using FileWriter::HDFFileBase::init;
+};
+
+class InMemoryHDFFile : public DebugHDFFile {
+public:
+  InMemoryHDFFile();
+};
+
+class DiskHDFFile : public DebugHDFFile {
+public:
+  DiskHDFFile(std::string const & FileName);
+};
+
+std::unique_ptr<DebugHDFFile> createInMemoryTestFile(const std::string &Filename,
                                            bool OnDisk = false);
 
 template <typename T>
